translated / renamed namespace
