+++ Spring 2024/Updated version of practice work/KloHarjoitusTyö/CoolApp/AddNewProduct.xaml.cs
@@ -12,7 +12,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
 
-namespace KloHarjoitusTyö
+namespace CoolApp
 {
     /// <summary>
     /// Interaction logic for AddNewProduct.xaml
