+++ Spring 2024/Updated version of practice work/KloHarjoitusTyö/CoolApp/AllInvoices.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="KloHarjoitusTyö.AllInvoices"
+﻿<Window x:Class="CoolApp.AllInvoices"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:KloHarjoitusTyö"
+        xmlns:local="clr-namespace:CoolApp"
         mc:Ignorable="d"
         Title="All Invoices" Height="400" Width="800">
     <Grid>
