+++ Spring 2024/Updated version of practice work/KloHarjoitusTyö/CoolApp/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="KloHarjoitusTyö.App"
+﻿<Application x:Class="CoolApp.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:KloHarjoitusTyö"
+             xmlns:local="clr-namespace:CoolApp"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
     </Application.Resources>
