+++ Spring 2024/Updated version of practice work/KloHarjoitusTyö/CoolApp/Invoice.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace KloHarjoitusTyö
+namespace CoolApp
 {
 
     /// <summary>
