+++ Spring 2024/Updated version of practice work/KloHarjoitusTyö/CoolApp/InvoiceLine.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace KloHarjoitusTyö
+namespace CoolApp
 {
     /// <summary>
     /// InvoiceLine luokka
