+++ Spring 2024/Updated version of practice work/KloHarjoitusTyö/CoolApp/MainWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="KloHarjoitusTyö.MainWindow"
+﻿<Window x:Class="CoolApp.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:KloHarjoitusTyö"
+        xmlns:local="clr-namespace:CoolApp"
         mc:Ignorable="d"
         Title="Invoicing" Height="350" Width="625" x:Name="mainWindow" ResizeMode="NoResize">
     <!--<Window.DataContext>
