+++ Spring 2024/Updated version of practice work/KloHarjoitusTyö/CoolApp/MainWindow.xaml.cs
@@ -14,7 +14,7 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace KloHarjoitusTyö
+namespace CoolApp
 {
     /// <summary>
     /// Interaction logic for MainWindow.xaml
