+++ Spring 2024/Updated version of practice work/KloHarjoitusTyö/CoolApp/NewInvoice.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="KloHarjoitusTyö.NewInvoice"
+﻿<Window x:Class="CoolApp.NewInvoice"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:KloHarjoitusTyö"
+        xmlns:local="clr-namespace:CoolApp"
         mc:Ignorable="d"
         Title="New Invoice" Height="450" Width="800">
     <Grid>
