+++ Spring 2024/Updated version of practice work/KloHarjoitusTyö/CoolApp/NewInvoice.xaml.cs
@@ -14,7 +14,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
 
-namespace KloHarjoitusTyö
+namespace CoolApp
 {
     public partial class NewInvoice : Window
