+++ Spring 2024/Updated version of practice work/KloHarjoitusTyö/CoolApp/Product.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace KloHarjoitusTyö
+namespace CoolApp
 {
     /// <summary>
     /// Product luokka
