+++ Spring 2024/Updated version of practice work/KloHarjoitusTyö/CoolApp/ShowAllProducts.xaml.cs
@@ -13,7 +13,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
 
-namespace KloHarjoitusTyö
+namespace CoolApp
 {
     /// <summary>
     /// Interaction logic for ShowAllProducts.xaml
